# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Vitas Project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/evaluates": {
            "get": {
                "tags": [
                    "Evaluates"
                ],
                "summary": "Get all evaluates",
                "description": "Get all evaluates",
                "operationId": "2f2c72633677dd3a08d5992479135e79",
                "responses": {
                    "200": {
                        "description": "Get Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Evaluates"
                ],
                "summary": "Create a new evaluate",
                "description": "Create a new evaluate",
                "operationId": "d081a11e518c2cc8fbe1cca8f0151df6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validated fail!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/evaluates/{id}": {
            "delete": {
                "tags": [
                    "Evaluates"
                ],
                "summary": "Delete evaluate by id",
                "description": "Delete evaluate by id",
                "operationId": "e97088859f5e64c725c318dcb8e72ee7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the evaluate",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Successfully!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}